Several basic tests now working
Including the one that made my life a living hell - check if card is locked. I start to feel like if it will be nice to throw some exceptions but that can be implemented and refactored later.

The problem was that I wasn't using the bankMock (before refactor - cardMock). So I assumed that the data will just fill itself. The teacher helped me with a tip and after some questions to chatgpt and looking at previous lecture I found out that the doAnswer was the right way to mock the isLocked behaviour.