Test if card is inserted
In case the card is inserted and valid there will be a flag that will be checked in each method so no need to always check if it is inserted in the other methods. It would have been the best if I could have listen if that boolean changes value and exit the process, but it is not required, so I won't be dealing with it.